requirements.txt

streamlit==1.31.0
langchain==0.1.6
langchain-community==0.0.19
langchain-google-genai==0.0.9
google-generativeai>=0.3.2
pypdf==3.17.4
python-dotenv==1.0.1
faiss-cpu==1.7.4


# streamlit==1.31.0
# langchain==0.1.6
# langchain-community==0.0.19
# langchain-google-genai==0.0.9
# google-generativeai>=0.3.2
# pypdf==3.17.4
# python-dotenv==1.0.1
# faiss-cpu==1.7.4
# numpy==1.26.0

